requirements.txt
discord.py
python-dotenv
selenium
PyYAML
webdriver-manager
